moved becs array into store, updated bec chips to properly handle text overflow
