+++ src/components/shared/BECChip.vue
@@ -1,22 +1,15 @@
 <template>
-  <v-chip :color="backgroundColor" :style="`color: ${color}`">{{ category }}</v-chip>
+  <v-chip
+    :color="backgroundColor"
+    :style="`color: ${color}; text-overflow: ellipsis; overflow: hidden;`"
+  >
+    <span class="text-truncate">{{ category }}</span>
+  </v-chip>
 </template>
 <script lang="ts">
 import { Vue, Component, Prop } from "vue-property-decorator";
 
-type BECCategory =
-  | "Superstructure"
-  | "Substructure"
-  | "Mechanical Services"
-  | "Electrical Services"
-  | "Public Health & Hydraulics"
-  | "Space plan"
-  | "Building Envelope";
-export interface BEC {
-  name: BECCategory;
-  color: string;
-  backgroundColor: string;
-}
+import { BEC, BECCategory } from "@/models/shared";
 
 @Component
 export default class BECChip extends Vue {
@@ -29,42 +22,8 @@ export default class BECChip extends Vue {
     return this.becs.find((b) => b.name === this.category)?.color;
   }
 
-  becs: BEC[] = [
-    {
-      name: "Superstructure",
-      color: "white",
-      backgroundColor: "#224a63",
-    },
-    {
-      name: "Substructure",
-      color: "black",
-      backgroundColor: "#aeebdb",
-    },
-    {
-      name: "Mechanical Services",
-      color: "black",
-      backgroundColor: "#f0b4b4",
-    },
-    {
-      name: "Electrical Services",
-      color: "white",
-      backgroundColor: "#754792",
-    },
-    {
-      name: "Public Health & Hydraulics",
-      color: "black",
-      backgroundColor: "#dbb5ea",
-    },
-    {
-      name: "Space plan",
-      color: "white",
-      backgroundColor: "#4b97d2",
-    },
-    {
-      name: "Building Envelope",
-      color: "black",
-      backgroundColor: "#82c7f1",
-    },
-  ];
+  get becs(): BEC[] {
+    return this.$store.state.becs;
+  }
 }
 </script>
