+++ src/views/Assessment.vue
@@ -18,6 +18,7 @@
           :totalVolume="totalVolume"
           :emptyProps="emptyProps"
           :report="report"
+          :becs="becs"
         />
       </v-col>
       <v-col cols="8">
@@ -108,6 +109,10 @@ export default class Assessment extends Vue {
       });
     });
     this.transportTypes = this.$store.state.transportTypes;
+  }
+
+  get becs(): string[] {
+    return this.$store.state.becs.map((b: { name: string }) => b.name);
   }
 
   async save() {
@@ -435,7 +440,6 @@ export default class Assessment extends Vue {
     this.projectData = data;
     this.$store.dispatch("changeRegion", data.region).then(res=>{
       this.materials = this.$store.getters.materialsArr
-      console.log(data.region)
     })
   }
 }
